WIP: refactoring, ln_platform
